+++ tests/conftest.py
@@ -26,6 +26,7 @@ RESOURCE_ID = "c4e3a9fb-4415-488e-ba57-d05269b27adf"
 RESOURCE_EXCEPTION_ID = "d4e3a9fb-4415-488e-ba57-d05269b27adf"
 RESOURCE_EXCEPTION_TABLE_INDEXES = {"Nom": "index", "N° de certificat": "index"}
 DATASET_ID = "601ddcfc85a59c3a45c2435a"
+NOT_EXISTING_RESOURCE_ID = "5d0b2b91-b21b-4120-83ef-83f818ba2451"
 pytestmark = pytest.mark.asyncio
 
 nest_asyncio.apply()