+++ udata_hydra/routes/__init__.py
@@ -15,6 +15,12 @@ from udata_hydra.routes.resources_exceptions import (
     delete_resource_exception,
     get_all_resources_exceptions,
 )
+from udata_hydra.routes.resources_legacy import (
+    create_resource_legacy,
+    delete_resource_legacy,
+    get_resource_legacy,
+    update_resource_legacy,
+)
 from udata_hydra.routes.status import get_crawler_status, get_health, get_stats, get_worker_status
 
 
@@ -47,14 +53,11 @@ routes_params = [
     (web.get, "/api/checks/all", get_all_checks, None),
     (web.post, "/api/checks", create_check, None),
     # Routes for resources
-    (web.get, "/api/resources", get_resource, None),
+    (web.get, "/api/resources/{resource_id}", get_resource, None),
     (web.get, "/api/resources/{resource_id}/status", get_resource_status, None),
     (web.post, "/api/resources", create_resource, None),
-    (web.put, "/api/resources", update_resource, None),
-    (web.delete, "/api/resources", delete_resource, None),
-    (web.post, "/api/resource/created", create_resource, None),  # TODO: legacy, to remove
-    (web.post, "/api/resource/updated", update_resource, None),  # TODO: legacy, to remove
-    (web.post, "/api/resource/deleted", delete_resource, None),  # TODO: legacy, to remove
+    (web.put, "/api/resources/{resource_id}", update_resource, None),
+    (web.delete, "/api/resources/{resource_id}", delete_resource, None),
     # Routes for statuses
     (web.get, "/api/status/crawler", get_crawler_status, None),
     (web.get, "/api/status/worker", get_worker_status, None),
@@ -73,3 +76,13 @@ routes_params = [
 
 # Generate the routes
 routes: list[web.RouteDef] = generate_routes(routes_params)
+
+# TODO: legacy, to remove
+legacy_routes_params = [
+    (web.get, "/api/resources", get_resource_legacy, None),
+    (web.post, "/api/resource/created", create_resource_legacy, None),
+    (web.post, "/api/resource/updated", update_resource_legacy, None),
+    (web.post, "/api/resource/deleted", delete_resource_legacy, None),
+]
+legacy_routes: list[web.RouteDef] = generate_routes(legacy_routes_params)
+routes.extend(legacy_routes)